requirements.txt
logzero
youtube-data-api
python-shout3
pafy
flask
flask-cors